refactor: move bool code
