+++ caste.go
@@ -7,59 +7,8 @@ package cast
 
 import (
 	"encoding/json"
-	"fmt"
 	"reflect"
-	"strconv"
-	"time"
 )
-
-// ToBoolE casts an interface to a bool type.
-func ToBoolE(i interface{}) (bool, error) {
-	i = indirect(i)
-
-	switch b := i.(type) {
-	case bool:
-		return b, nil
-	case nil:
-		return false, nil
-	case int:
-		return b != 0, nil
-	case int64:
-		return b != 0, nil
-	case int32:
-		return b != 0, nil
-	case int16:
-		return b != 0, nil
-	case int8:
-		return b != 0, nil
-	case uint:
-		return b != 0, nil
-	case uint64:
-		return b != 0, nil
-	case uint32:
-		return b != 0, nil
-	case uint16:
-		return b != 0, nil
-	case uint8:
-		return b != 0, nil
-	case float64:
-		return b != 0, nil
-	case float32:
-		return b != 0, nil
-	case time.Duration:
-		return b != 0, nil
-	case string:
-		return strconv.ParseBool(i.(string))
-	case json.Number:
-		v, err := ToInt64E(b)
-		if err == nil {
-			return v != 0, nil
-		}
-		return false, fmt.Errorf("unable to cast %#v of type %T to bool", i, i)
-	default:
-		return false, fmt.Errorf("unable to cast %#v of type %T to bool", i, i)
-	}
-}
 
 // From html/template/content.go
 // Copyright 2011 The Go Authors. All rights reserved.
